8629: move provider names logic to vaos
